Removing code that causes compiler warnings
There were several places where a halt instruction was being
used after a while(1) loop. These halt instructions have
been removed.

There were also two instances where a 32 bit pointer was
getting used as an 8 bit pointer which caused a compiler
warning. This has been corrected.

+++ examples/am335x/PRU_access_const_table/PRU_access_const_table.c
@@ -53,7 +53,7 @@ PRU_SRAM volatile uint32_t shared_freq_2;
 PRU_SRAM volatile uint32_t shared_freq_3;
 
 /* PRCM Registers */
-#define CM_PER_BASE	((volatile uint32_t *)(0x44E00000))
+#define CM_PER_BASE	((volatile uint8_t *)(0x44E00000))
 #define SPI0_CLKCTRL  (0x4C)
 #define ON (0x2)
